Move array list record mapper to array list package
